chore: new file for visualforceutils, refactoring
