+++ test/utilities/index.ts
@@ -15,3 +15,4 @@ export * from './outputView';
 export * from './sideBar';
 export * from './statusBar';
 export * from './terminalView';
+export * from './visualforceUtils';
